feat: restructure commit
