+++ src/maze_visualization/__init__.py
@@ -1,0 +1,4 @@
+"""Maze Game graphics definition."""
+from src.maze_visualization.maze_game_visualization import MazeGameVisualization
+
+__all__ = ["MazeGameVisualization"]
