+++ src/maze_visualization/game_color.py
